refactor: static 메소드 제거, LottoTicketList 클래스 내 함수 재정의

+++ src/main/java/view/OutputView.java
@@ -31,7 +31,7 @@ public class OutputView {
         printProfitRate(lottoResults.getProfitRate());
     }
 
-    private static void printProfitRate(float profitRate) {
+    private static void printProfitRate(double profitRate) {
         if (profitRate < 1.0) {
             System.out.printf("총 수익률은 %.2f입니다.(기준이 1이기 때문에 결과적으로 손해라는 의미임)\n", profitRate);
             return;
